Create m.file
In this code, we first set the sampling frequency Fs (here assumed to be 1000 Hz). Then, we set the cutoff frequency of the high-pass filter Fc to 50 MHz. Next, we specify the filter order N (here assumed to be 4). We design the high-pass filter using the butter function. Finally, we plot the frequency response of the filter using the freqz function.